requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
pdf2image==1.17.0
python-multipart==0.0.9
pydantic==2.6.4
numpy==1.26.4
google-genai==1.3.0
